feat: Refactor DNS management to use AWS Lambda and API Gateway, removing previous server implementation
