--- amplify.yml
@@ -1,34 +1,0 @@
-version: 1
-applications:
-  - frontend:
-      phases:
-        preBuild:
-          commands:
-            - cd server && npm install
-            - cd .. && npm install
-        build:
-          commands:
-            - npm run build
-      artifacts:
-        baseDirectory: build
-        files:
-          - '**/*'
-      cache:
-        paths:
-          - node_modules/**/*
-          - server/node_modules/**/*
-    backend:
-      phases:
-        preBuild:
-          commands:
-            - cd server && npm install
-        build:
-          commands:
-            - cd server && npm run start &
-      artifacts:
-        baseDirectory: server
-        files:
-          - '**/*'
-      cache:
-        paths:
-          - server/node_modules/**/*
